requirements.txt
flask
torch
torchvision
pillow
numpy
werkzeug
flask-cors
